Split dpdk-rdk recipe and move rdk-user to it's own recipe
A better implementation for dpdk and rdk-userspace is to split
dpdk-rdk recipe and move rdk-userspace to its own recipe.

rdk-user will be a build-time dependency for dpdk-rdk and all the
libraries and headers will be located in dpdk-rdk recipe sysroot,
so it will be possible to build dpdk with the RDK patch applied.

Also, update axxia-image-[large|sim] to with the new package name.

Signed-off-by: Daniel Dragomir <daniel.dragomir@windriver.com>