dpdk-rdk: Build shared instead static libraries
Enable CONFIG_RTE_BUILD_SHARED_LIB option in the defconfig
for RTE_TARGET associated with x86-64 arch to build shared
libraries instead static ones.

Enable some other options which are enabled from Makefiles
used for standalone build (not in Yocto build system). One
of them is CONFIG_RTE_EAL_NUMA_AWARE_HUGEPAGES which require
enabling numa support (with PACKAGECONFIG mechanism).

Also, disable devel build to avoid rpath in binaries. By
default, development build is enabled if DPDK sources are
in a git tree.

Also, install dpdk-pmdinfogen binary from app sub-directory
which is skiped untill now because it was installed in a
wrong location, not in ${bindir}.

If shared libraries are chosen, in examples sub-directory a
new library is build for ethtool example wich is not installed
in the default (makefile) procedure. Install it in the main
sub-package and the symlink in -dev sub-package and make them
run-time dependencies for ${PN}-examples sub-package.

A0228 - dpdk: static and shared libraries

Signed-off-by: Daniel Dragomir <daniel.dragomir@windriver.com>