linux: Customize the RDK Staging Directory
Add the ability to set the version of the RDK KLM modules throught
RDK_KLM_VERSION variable in order to customize the path where the
drivers will be installed in the kernel tree
(drivers/staging/intel/<rdk-klm-version>) and to customize the commit
message in the kernel git repository from work-shared directory.

If this variable is not set in local.conf, its value will be extracted
from the RDK KLM archive name (rdk_klm_src_<releaseinfo>.txz).

A0269 - Control the RDK Staging Directory

Signed-off-by: Daniel Dragomir <daniel.dragomir@windriver.com>