linux: Refactoring do_patch_rdk_modules code
For easier reading the code, rename DEFAULT_PATH variable to
STAGING_PATH and concatenate it later with the RDK_KLM_VERSION.

Also add a suplimentar check if RELEASE_TAG extracted from the
archive name is empty and set it to "unknown_release_info".
This case shouldn't be valid if the archives are named as before.

A0269 - Control the RDK Staging Directory

Signed-off-by: Daniel Dragomir <daniel.dragomir@windriver.com>