linux: Fix kernel options warnings
Get rid of all the build warnings related to kernel options
mismatches or not defined in the kernel tree.

The cause for those warnings is that for some kernel options
we set a different value in meta-intel-axxia than defined upstream
(in other fragments from other meta layers or yocto-kernel-cache).
Bitbake is not complainig about overwritten kernel options, but it
does if those kernel options are dependencies for other options
which disapear or are unset in the final config because of their
dependenies even if they are explicitly enabled in fragments.

A0259 - snr: Lots of kernel option warnings on thud branch

Signed-off-by: Daniel Dragomir <daniel.dragomir@windriver.com>