rdk-tools: Recreate symlinks removed on install
THe install target from the main Makefile of the package creates
a install directory and copy there all the files which should be
in rootfs. But for the copying proces it use rsync with the following
parameters:

RSYNC_CMD="rsync -aL --no-l" ;

-L and --no-l flags "transform symlink into referent file/dir".

Correct this in do_install function in the recipe.

B0168 - libies_sdk-50sm.so.2 packaging problem

Signed-off-by: Daniel Dragomir <Daniel.Dragomir@windriver.com>