rdk-tools: RPC mode compile fix
The host contaminations cames from ax_check_openssl.m4 lib which
search for openssl. If the openssl location is not set via
--with-openssl on configure, or via OPENSSL_ROOT env variable or
via pkg-config tool, the build system will search on host.

If openssl libs are found on host, -I/usr/include path will be
added as build flag with lead to host contamination:
cc1plus: warning: include location "/usr/include" is unsafe
	for cross-compilation [-Wpoison-system-directories]

If host path is added in build frags, it will search for each
header in the host first, and if host headers are old or
imcompatible with the rdk-tools version, various error will
appear. This is why with Ubuntu 18.04 the build works, but
with 12.04 it fails (features.h header is the cause for the
first errors from teh log).

Signed-off-by: Daniel Dragomir <daniel.dragomir@windriver.com>